+++ .idea/gradle.xml
@@ -7,7 +7,7 @@
         <compositeConfiguration>
           <compositeBuild compositeDefinitionSource="SCRIPT">
             <builds>
-              <build path="$PROJECT_DIR$/plugin-build" name="io.frankois944.spmForKmp">
+              <build path="$PROJECT_DIR$/plugin-build" name="io.github.frankois944.spmForKmp">
                 <projects>
                   <project path="$PROJECT_DIR$/plugin-build" />
                   <project path="$PROJECT_DIR$/plugin-build/plugin" />
