+++ plugin-build/gradle.properties
@@ -1,13 +1,8 @@
-ID=io.frankois944.spmForKmp.plugin
+ID=io.github.frankois944.spmForKmp.plugin
 VERSION=0.0.1
-GROUP=io.frankois944.spmForKmp
+GROUP=io.github.frankois944.spmForKmp
 DISPLAY_NAME=Swift Package Manager for Kotlin Multiplaform
 DESCRIPTION=A Gradle plugin for integrating Swift packages into a Kotlin multiplatform module
 WEBSITE=https://github.com/frankois944/spm4Kmp
 VCS_URL=https://github.com/frankois944/spm4Kmp
 IMPLEMENTATION_CLASS=io.frankois944.spmForKmp.plugin.SpmForKmpPlugin
-mavenCentralUsername = qOiRXw6p
-mavenCentralPassword = a7fbQmDzH6AdXk6T5L174c2o5s2gggsbpRTF5bQ36PED
-signing.keyId = F41CF3FE
-signing.password = Soleil11$
-signing.secretKeyRingFile = /Users/francoisdabonot/DEV/secret_key.gpg
