--- plugin-build/plugin/src/functionalTest/kotlin/io/frankois944/spmForKmp/plugin/BasicPackageTest.kt
@@ -1,43 +1,0 @@
-package io.frankois944.spmForKmp.plugin
-
-import com.autonomousapps.kit.GradleBuilder.build
-import com.autonomousapps.kit.truth.TestKitTruth.Companion.assertThat
-import io.frankois944.spmForKmp.plugin.fixture.KotlinSource
-import io.frankois944.spmForKmp.plugin.fixture.SmpKMPTestFixture
-import io.frankois944.spmForKmp.plugin.fixture.SwiftSource
-import org.junit.jupiter.api.Test
-
-class BasicPackageTest {
-    @Test
-    fun `build with imported UIKit framework is successful`() {
-        // Given
-        val fixture =
-            SmpKMPTestFixture
-                .builder()
-                .withSwiftSources(
-                    SwiftSource.of(
-                        content =
-                            """
-                            import UIKit
-                            @objc public class TestView: UIView {}
-                            """.trimIndent(),
-                    ),
-                ).withKotlinSources(
-                    KotlinSource.of(
-                        content =
-                            """
-                            package com.example
-                            import dummy.TestView
-                            @kotlinx.cinterop.ExperimentalForeignApi
-                            val view = TestView()
-                            """.trimIndent(),
-                    ),
-                ).build()
-
-        // When
-        val result = build(fixture.gradleProject.rootDir, "build")
-
-        // Then
-        assertThat(result).task(":library:build").succeeded()
-    }
-}
