+++ plugin-build/plugin/src/functionalTest/kotlin/io/github/frankois944/spmForKmp/plugin/RemotePackageTest.kt
@@ -1,0 +1,236 @@
+package io.github.frankois944.spmForKmp.plugin
+
+import com.autonomousapps.kit.GradleBuilder.build
+import com.autonomousapps.kit.truth.TestKitTruth.Companion.assertThat
+import io.github.frankois944.spmForKmp.plugin.definition.SwiftDependency
+import io.github.frankois944.spmForKmp.plugin.fixture.KotlinSource
+import io.github.frankois944.spmForKmp.plugin.fixture.SmpKMPTestFixture
+import io.github.frankois944.spmForKmp.plugin.fixture.SwiftSource
+import io.github.frankois944.spmForKmp.plugin.utils.OpenFolderOnFailureExtension
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.RegisterExtension
+
+class RemotePackageTest {
+    private var folderTopOpen: String? = null
+
+    @RegisterExtension
+    @JvmField
+    val openFolderOnFailure =
+        OpenFolderOnFailureExtension {
+            folderTopOpen ?: ""
+        }
+
+    @BeforeEach
+    fun beforeEach() {
+        folderTopOpen = null
+    }
+
+    @Test
+    fun `build with remote packages by version`() {
+        // Given
+        val fixture =
+            SmpKMPTestFixture
+                .builder()
+                .withDependencies(
+                    buildList {
+                        add(
+                            SwiftDependency.Package.Remote.Version(
+                                url = "https://github.com/krzyzanowskim/CryptoSwift.git",
+                                names = listOf("CryptoSwift"),
+                                version = "1.8.3",
+                                exportToKotlin = true,
+                            ),
+                        )
+                    },
+                ).withKotlinSources(
+                    KotlinSource.of(
+                        content =
+                            """
+                            package com.example
+                            import CryptoSwift.SWIFT_TYPEDEFS
+                            import dummy.MySwiftClass
+                            """.trimIndent(),
+                    ),
+                ).withSwiftSources(
+                    SwiftSource.of(
+                        content =
+                            """
+                            import Foundation
+                            import CryptoSwift
+                            @objc public class MySwiftClass: NSObject {
+                                @objc public func toMD5(value: String) -> String {
+                                    return value.md5()
+                                }
+                            }
+                            """.trimIndent(),
+                    ),
+                ).build()
+
+        val project = fixture.gradleProject.rootDir
+        folderTopOpen = project.absolutePath
+        // When
+        val result =
+            build(project, "build")
+
+        // Then
+        assertThat(result).task(":library:build").succeeded()
+    }
+
+    @Test
+    fun `build with remote packages by branch`() {
+        // Given
+        val fixture =
+            SmpKMPTestFixture
+                .builder()
+                .withDependencies(
+                    buildList {
+                        add(
+                            SwiftDependency.Package.Remote.Branch(
+                                url = "https://github.com/krzyzanowskim/CryptoSwift.git",
+                                names = listOf("CryptoSwift"),
+                                packageName = "CryptoSwift",
+                                branch = "main",
+                                exportToKotlin = true,
+                            ),
+                        )
+                    },
+                ).withKotlinSources(
+                    KotlinSource.of(
+                        content =
+                            """
+                            package com.example
+                            import CryptoSwift.SWIFT_TYPEDEFS
+                            import dummy.MySwiftClass
+                            """.trimIndent(),
+                    ),
+                ).withSwiftSources(
+                    SwiftSource.of(
+                        content =
+                            """
+                            import Foundation
+                            import CryptoSwift
+                            @objc public class MySwiftClass: NSObject {
+                                @objc public func toMD5(value: String) -> String {
+                                    return value.md5()
+                                }
+                            }
+                            """.trimIndent(),
+                    ),
+                ).build()
+
+        val project = fixture.gradleProject.rootDir
+        folderTopOpen = project.absolutePath
+        // When
+        val result =
+            build(project, "build")
+
+        // Then
+        assertThat(result).task(":library:build").succeeded()
+    }
+
+    @Test
+    fun `build with remote packages by commit`() {
+        // Given
+        val fixture =
+            SmpKMPTestFixture
+                .builder()
+                .withDependencies(
+                    buildList {
+                        add(
+                            SwiftDependency.Package.Remote.Commit(
+                                url = "https://github.com/krzyzanowskim/CryptoSwift.git",
+                                names = listOf("CryptoSwift"),
+                                revision = "729e01bc9b9dab466ac85f21fb9ee2bc1c61b258",
+                                exportToKotlin = true,
+                            ),
+                        )
+                    },
+                ).withKotlinSources(
+                    KotlinSource.of(
+                        content =
+                            """
+                            package com.example
+                            import CryptoSwift.SWIFT_TYPEDEFS
+                            import dummy.MySwiftClass
+                            """.trimIndent(),
+                    ),
+                ).withSwiftSources(
+                    SwiftSource.of(
+                        content =
+                            """
+                            import Foundation
+                            import CryptoSwift
+                            @objc public class MySwiftClass: NSObject {
+                                @objc public func toMD5(value: String) -> String {
+                                    return value.md5()
+                                }
+                            }
+                            """.trimIndent(),
+                    ),
+                ).build()
+
+        val project = fixture.gradleProject.rootDir
+        folderTopOpen = project.absolutePath
+        // When
+        val result =
+            build(project, "build")
+
+        // Then
+        assertThat(result).task(":library:build").succeeded()
+    }
+
+    @Test
+    fun `build with complex remote packages`() {
+        // Given
+        val fixture =
+            SmpKMPTestFixture
+                .builder()
+                .withTargets(CompileTarget.iosSimulatorArm64, CompileTarget.iosArm64)
+                .withDependencies(
+                    buildList {
+                        add(
+                            SwiftDependency.Package.Remote.Version(
+                                url = "https://github.com/firebase/firebase-ios-sdk.git",
+                                names = listOf("FirebaseCore", "FirebaseAnalytics", "FirebaseCrashlytics"),
+                                version = "11.6.0",
+                                packageName = "firebase-ios-sdk",
+                                exportToKotlin = true,
+                            ),
+                        )
+                    },
+                ).withKotlinSources(
+                    KotlinSource.of(
+                        content =
+                            """
+                            package com.example
+                            import FirebaseCore.FIRApp
+                            import FirebaseCrashlytics.FIRCrashlyticsMeta
+                            import FirebaseAnalytics.FIRConsentStatusGranted
+                            """.trimIndent(),
+                    ),
+                ).withSwiftSources(
+                    SwiftSource.of(
+                        content =
+                            """
+                            import Foundation
+                            import FirebaseCore
+                            import FirebaseAnalytics
+                            import FirebaseCrashlytics
+
+                            @objc public class MySwiftClass: NSObject {
+                            }
+                            """.trimIndent(),
+                    ),
+                ).build()
+
+        val project = fixture.gradleProject.rootDir
+        folderTopOpen = project.absolutePath
+        // When
+        val result =
+            build(project, "build")
+
+        // Then
+        assertThat(result).task(":library:build").succeeded()
+    }
+}
